+++ waspc/cli/src/Wasp/Cli/Command/BashCompletion.hs
@@ -44,7 +44,6 @@ bashCompletion = do
         "telemetry",
         "deps",
         "dockerfile",
-        "info",
         "show",
         "test",
         "studio"